# Using tshark JSON output to speed up pyshark