--- a/src/pyshark/capture/live_capture.py
+++ b/src/pyshark/capture/live_capture.py
@@ -6,6 +6,7 @@ import sys
 if sys.version_info >= (3, 0):
     basestring = str
 
+
 class LiveCapture(Capture):
     """
     Represents a live capture on a network interface.
@@ -13,7 +14,7 @@ class LiveCapture(Capture):
 
     def __init__(self, interface=None, bpf_filter=None, display_filter=None, only_summaries=False, decryption_key=None,
                  encryption_type='wpa-pwk', output_file=None, decode_as=None, disable_protocol=None, tshark_path=None,
-                 override_prefs=None, capture_filter=None, monitor_mode=None):
+                 override_prefs=None, capture_filter=None, monitor_mode=None, use_json=False):
         """
         Creates a new live capturer on a given interface. Does not start the actual capture itself.
 
@@ -32,12 +33,14 @@ class LiveCapture(Capture):
         :param override_prefs: A dictionary of tshark preferences to override, {PREFERENCE_NAME: PREFERENCE_VALUE, ...}.
         :param capture_filter: Capture (wireshark) filter to use.
         :param disable_protocol: Tells tshark to remove a dissector for a specifc protocol.
+        :param use_json: Uses tshark in JSON mode (EXPERIMENTAL). It is a good deal faster than XML
+        but also has less information. Available from Wireshark 2.2.0.
         """
         super(LiveCapture, self).__init__(display_filter=display_filter, only_summaries=only_summaries,
                                           decryption_key=decryption_key, encryption_type=encryption_type,
                                           output_file=output_file, decode_as=decode_as, disable_protocol=disable_protocol,
                                           tshark_path=tshark_path, override_prefs=override_prefs,
-                                          capture_filter=capture_filter)
+                                          capture_filter=capture_filter, use_json=use_json)
         self.bpf_filter = bpf_filter
         self.monitor_mode = monitor_mode