--- a/src/pyshark/packet/fields.py
+++ b/src/pyshark/packet/fields.py
@@ -1,0 +1,114 @@
+import binascii
+
+from pyshark.packet.common import Pickleable
+
+
+class LayerField(object):
+    """
+    Holds all data about a field of a layer, both its actual value and its name and nice representation.
+    """
+    # Note: We use this object with slots and not just a dict because
+    # it's much more memory-efficient (cuts about a third of the memory).
+    __slots__ = ['name', 'showname', 'raw_value', 'show', 'hide', 'pos', 'size', 'unmaskedvalue']
+
+    def __init__(self, name=None, showname=None, value=None, show=None, hide=None, pos=None, size=None, unmaskedvalue=None):
+        self.name = name
+        self.showname = showname
+        self.raw_value = value
+        self.show = show
+        self.pos = pos
+        self.size = size
+        self.unmaskedvalue = unmaskedvalue
+
+        if hide and hide == 'yes':
+            self.hide = True
+        else:
+            self.hide = False
+
+    def __repr__(self):
+        return '<LayerField %s: %s>' % (self.name, self.get_default_value())
+
+    def get_default_value(self):
+        """
+        Gets the best 'value' string this field has.
+        """
+        val = self.show
+        if not val:
+            val = self.raw_value
+        if not val:
+            val = self.showname
+        return val
+
+    @property
+    def showname_value(self):
+        """
+        For fields which do not contain a normal value, we attempt to take their value from the showname.
+        """
+        if self.showname and ': ' in self.showname:
+            return self.showname.split(': ')[1]
+
+    @property
+    def showname_key(self):
+        if self.showname and ': ' in self.showname:
+            return self.showname.split(': ')[0]
+
+    def __getstate__(self):
+        ret = {}
+        for slot in self.__slots__:
+            ret[slot] = getattr(self, slot)
+        return ret
+
+    def __setstate__(self, data):
+        for key, val in data.iteritems():
+            setattr(self, key, val)
+
+    @property
+    def binary_value(self):
+        """
+        Returns the raw value of this field (as a binary string)
+        """
+        return binascii.unhexlify(self.raw_value)
+
+    @property
+    def int_value(self):
+        """
+        Returns the raw value of this field (as an integer).
+        """
+        return int(self.raw_value, 16)
+
+
+class LayerFieldsContainer(str, Pickleable):
+    """
+    An object which contains one or more fields (of the same name).
+    When accessing member, such as showname, raw_value, etc. the appropriate member of the main (first) field saved
+    in this container will be shown.
+    """
+
+    def __new__(cls, main_field, *args, **kwargs):
+        obj = str.__new__(cls, main_field.get_default_value(), *args, **kwargs)
+        obj.fields = [main_field]
+        return obj
+
+    def add_field(self, field):
+        self.fields.append(field)
+
+    @property
+    def main_field(self):
+        return self.fields[0]
+
+    @property
+    def alternate_fields(self):
+        """
+        Return the alternate values of this field containers (non-main ones).
+        """
+        return self.fields[1:]
+
+    @property
+    def all_fields(self):
+        """
+        Returns all fields in a list, the main field followed by the alternate fields.
+        """
+        return self.fields
+
+    def __getattr__(self, item):
+        return getattr(self.main_field, item)