This branch adds an experimental capacity to use JSON instead of XML parsing with tshark. This is a feature from Wireshark 2.2.0. It is not the default option.
Please review and make suggestions.

The speed-up is considerable, however multiple things should be noted:

- The JSON does not contain the same amount of information the XML does. The actual value of the field is present but the "nice" value and presentation are not. As such, I recommend it for programmatic use but not for interactive use (where performance is not important anyway).
- Field names may be slightly different, and in particular I use a different nesting for tree fields, which is to say that instead of accessing an inner field in a flat way like `pkt.tcp.flags_ack` you will access it via `pkt.tcp.flags.ack`. 
- Because of those two changes, existing pyshark code is not immediately portable to the JSON version.
- I have not comprehensively tested it, and it is not tested for InMem capture at all. 

Performance (in all tests packets are immediately discarded after parsing):

**Medium pcap (17MB mostly UDP, 20000 packets)**
Pyshark using XML: ~29.2 s
Pyshark using JSON: ~3.49 s
Scapy: ~7.08 s

**Large pcap (133M, ~150000 packets)**
Pyshark using XML: 3 min 43 s
Pyshark using JSON: 25.4 s
Scapy: Crashes :(